requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
PyJWT==2.8.0
bcrypt==4.1.2
Pillow==10.2.0
cloudinary==1.37.0
pymongo==4.6.1
djongo==1.3.6
django-extensions==3.2.3
python-dotenv==1.0.0